src/description-to-json/requirements.txt
openai
google-cloud-secretmanager